0822 J

+++ tracerProject/src/main/java/com/web/tracerProject/vo/ResourceManage.java
@@ -1,6 +1,5 @@
 package com.web.tracerProject.vo;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import org.springframework.format.annotation.DateTimeFormat;
 import lombok.AllArgsConstructor;
@@ -15,20 +14,20 @@ import lombok.Setter;
 public class ResourceManage {
     private String rid;
     private String rtype;
-    private BigDecimal assigned_budget;
-    private BigDecimal used_budget;
-    private BigDecimal remainingBudget;
+    private int assigned_budget;
+    private int used_budget;
+    private int remainingBudget;
     private String software_name;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date license_purchase_date;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date license_expiry_date;
-    private BigDecimal software_price;
+    private int software_price;
     private String pid;
     private String projectTitle;
 
     // 필요한 생성자 추가
-    public ResourceManage(String rid, String pid, String rtype, String software_name, Date license_purchase_date, Date license_expiry_date, BigDecimal software_price) {
+    public ResourceManage(String rid, String pid, String rtype, String software_name, Date license_purchase_date, Date license_expiry_date, int software_price) {
         this.rid = rid;
         this.pid = pid;
         this.rtype = rtype;
